+++ examples/react-webpack/rsbuild.config.ts
@@ -1,9 +1,0 @@
-import { defineConfig } from '@rsbuild/core';
-import { pluginReact } from '@rsbuild/plugin-react';
-import { pluginCssMinimizer } from '@rsbuild/plugin-css-minimizer';
-import { webpackProvider } from '@rsbuild/webpack';
-
-export default defineConfig({
-  plugins: [pluginReact(), pluginCssMinimizer()],
-  provider: webpackProvider,
-});