+++ packages/monorepo-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rsbuild/monorepo-utils
 
+## 0.2.7
+
+### Patch Changes
+
+- @rsbuild/shared@0.2.7
+
 ## 0.2.6
 
 ### Patch Changes