+++ packages/plugin-stylus/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rsbuild/plugin-stylus
 
+## 0.2.7
+
+### Patch Changes
+
+- @rsbuild/shared@0.2.7
+
 ## 0.2.6
 
 ### Patch Changes